Move prefix argument implementation to PrefixArgumentHandler class
